# combined reader gallop mode