In case when a single reader contributes a stream of fragments and keeps winning over other readers, mutation_reader_merger will enter gallop mode, in which it is assumed that the reader will keep winning over other readers. Currently, a reader needs to contribute 3 fragments to enter that mode.

In gallop mode, fragments returned by the galloping reader will be compared with the best fragment from _reader_heap or _fragment_heap. If it wins, the fragment is directly returned. Otherwise, gallop mode ends and merging performed as in general case, which involves heap operations.

A microbenchmark was added in order to test performance of the galloping reader optimization. A combining reader that merges results from four other readers is created. Each sub-reader provides a range of 16 clustering rows that is disjoint from others. All sub-readers return rows from the same partition. An improvement can be observed after introducing the galloping reader optimization.

As for other benchmarks from the "combined" group, results are pretty close to the old ones. Usually new results are slightly slower, but it depends on the run - in the comparision below, most results for the optimized version are actually better. Most of the time, the difference is not greater than 6%.
    
Median times from a single run of perf_mutation_readers.combined:
(1s run duration, 5 runs per benchmark, dev mode)

```    
test name                      before    -> after     change
combined.one_row:              103.416ns -> 105.080ns (1.6% worse)
combined.single_active:        101.730us -> 98.141us  (3.5% better)
combined.many_overlapping:     817.900us -> 808.799us (1.1% better)
combined.disjoint_interleaved: 96.008us  -> 92.657us  (3.4% better)
combined.disjoint_ranges:      96.100us  -> 91.972us  (4.3% better)
combined.overlapping_partitions_disjoint_rows:
                               107.784us -> 71.915us  (33.3% better)
```

Tested on AMD Ryzen Threadripper 2950X @ 3.5GHz.

Tests: unit(dev)
Fixes #3593.