@avikivity 2019-10-21
Please provide measurements in release mode, and normalize to time per fragment.