@avikivity 2019-10-22
@piodul does this work for different rows in the same partition, different partitions, or both?