@piodul 2019-10-22
@avikivity Currently, this optimization works only for different rows in the same partition. Galloping is turned off when next partition starts: https://github.com/scylladb/scylla/pull/5198/files#diff-2a2d8752bb687a0e338dfe715fcc3bd3R365