@piodul 2019-10-23
@denesb

I tried reorganizing the logic in the way you described (removing single and galloping readers, passing through `_new_fragments`), but observed a performance regression on the perf_mutation_readers microbenchmark.

Measurements from the first implementation of this change (these were run in release mode, on a slightly faster machine):

```
test                                      iterations      median         mad         min         max
combined.one_row                             2510366    51.178ns     0.179ns    50.985ns    51.607ns
combined.single_active                         10745    57.073us    77.093ns    56.830us    57.399us
combined.many_overlapping                       1601   477.804us     1.320us   476.484us   482.929us
combined.disjoint_interleaved                  10408    55.716us   230.158ns    55.284us    55.950us
combined.disjoint_ranges                       10352    54.414us    34.435ns    54.380us    54.739us
combined.overlapping_partitions_disjoint_rows       15147    38.201us    33.760ns    38.137us    38.544us
```

After changing the logic [link](https://github.com/piodul/scylla/compare/765dc86de43730b70cca993b2e8dcf78be424e28..9d723ebc77570ef475cbcfe2551217bf746819ea):

```
test                                      iterations      median         mad         min         max
combined.one_row                             2307771    56.470ns     0.126ns    56.134ns    56.596ns
combined.single_active                          8549    71.943us   582.406ns    70.296us    75.426us
combined.many_overlapping                       1591   476.175us   266.926ns   475.885us   485.330us
combined.disjoint_interleaved                   8662    74.187us   140.655ns    74.046us    74.480us
combined.disjoint_ranges                        8711    70.226us   214.404ns    70.012us    71.927us
combined.overlapping_partitions_disjoint_rows       12018    56.130us   382.621ns    55.734us    56.913us
```

[Reintroducing `_single_reader` special handling](https://github.com/piodul/scylla/commit/aa07f39bf7bddd1c8cd3d875b79af740b4eea590) helped bring back old performance for the half of the tests, but the new benchmark that tests the `galloping reader` optimization target is still the same:

```
test                                      iterations      median         mad         min         max
combined.one_row                             2435163    51.774ns     0.396ns    50.907ns    52.169ns
combined.single_active                         10734    56.813us     8.027ns    56.789us    56.910us
combined.many_overlapping                       1599   475.385us     1.840us   472.539us   479.347us
combined.disjoint_interleaved                  10250    55.051us   205.976ns    54.712us    57.079us
combined.disjoint_ranges                        9693    57.982us   345.795ns    57.221us    63.638us
combined.overlapping_partitions_disjoint_rows       11485    58.646us   263.370ns    58.367us    72.713us
```

I agree that clarity is important, but the goal of this issue was to optimize a special case. I'm not sure if introducing performance regression would count as a valid solution...

I have reorganized the code a little differently while keeping the performance (only `many_overlapping` benchmark time increased from 475us to 500us). It reduced code duplication, but ultimately I'm not sure it is now more clear or not. This version is available here: [link](https://github.com/piodul/scylla/commit/1f47d4f231c4e892c0e2f045109225b187ffe970). Which version would you prefer me to keep?

By the way, could you clarify what you meant by macrobenchmarks? I'm not sure what you meant, so I sticked only to running microbenchmarks.