@piodul 2019-10-29
Increased the `overlapping_partitions_disjoint_rows` to have 32 clustering rows per partition, in order to make the total amount of clustering rows used in the test similar to other ones (all benchmarks apart from `one_row` and `many_overlapping` have 512 clustering rows).

@denesb
As expected, changing compile-time parameter to dynamic one did not change the performance of the solution.

I tried running benchmarks from `perf_fast_forward`, but observed no difference in performance.

I also tried to see if a difference can be observed in real queries. I prepared a table with only one partition and 4 sstables. Each sstable had a range of 10*1024 rows, and each range was disjoint from other sstables' ranges. I did not observe any noticeable improvement when performing a `select * from test_table` query.

@avikivity
Benchmarks redone in release mode. New results are included in the message for commit with benchmark, and in the PR description.

As for preserving the gallop mode across partitions, should we implement it in this version, or in a follow-up?