@avikivity 2019-10-29
@piodul your select benchmark was perhaps dominated by paging latency. You need to saturate the server to overserve the throughput improvement. Also, did you prevent compaction from merging your tables?

The gain is about 90ns/row. We've seen nodes doing 200k rows/sec/core, which is about 5000 ns/row. So the improvement should be around 2%. But I expect the benefit to come for compactions when merging sstables that have disjoint data (which can be quite common for time series workloads as the data is ordered by timestamp).