@piodul 2019-10-29
@avikivity The table used TWCS strategy, I created each sstables in different time window, so they would not be compacted. The sstable files did not change before, during or after the test.