@avikivity 2019-10-29
I see. They may have been cached after the first read.

Maybe you can eliminate paging latency by using aggregation: select count(*) from tab where pk = ?.