@piodul 2019-10-30
@avikivity I re-did the query benchmark with following changes:
- disabled cache by `--enable-cache=no`
- used a larger table (16 sstables, 100*1024 clustering rows each, everything is still in the same partition)
- ran query `select (*) from test_table where pk = 0`

I actually managed to observe an improvement:

```
test           median     min        max        mad
before change  2.171087s  2.166947s  2.222554s  0.002911s
after change   1.870574s  1.817358s  1.917264s  0.007278s
```

median difference: 0.300513s (13.84%)